Pirates Module - work in progress
Spawn clusters of pirates whom either have an ambush behaviour or a wingman behavior, following one who is ambushing.

TODO: other behaviors that AI can take on.
TODO: the choice to attack or not function needs to be something that can be saved - so data driven.
TODO: the the whole thing needs to be able to save/load with the game so behaviors persist across games.